se agrego estilos a las tablas y corriguio algunos errores de gestion de cursos

+++ adm/administrar_coferencias.php
@@ -13,14 +13,19 @@ require_once '../conexion.php';
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="../style/style.css">
+    <link rel="stylesheet" href="../style/style1.css">
     <style>
-        body{
+        h1{
+            color: #A65B1A;
+        }
+        /* body{
             background-color: #333333;
             color: #f5f5f5;
         }
         center{
             padding-top: 5%;
-        }
+        } */
+        
     </style>
     <title>Documento</title>
 </head>
@@ -28,32 +33,33 @@ require_once '../conexion.php';
     
 <center>
         <form method="post" action="../controller/crud_cursos/insertar.php"> 
-            <h2 class="titulo_f">Nuevo Curso</h2>
+            <br><h1 class="titulo_f">Nuevo Curso</h1>
             <br>
             <input type="text" placeholder="nombre del curso" required name="nombre_curso"><br>
             <input type="text" placeholder="expositor" required name="expositor"><br>
-            <textarea  required name="comentario">descripcion </textarea><br> 
             <input type="int" placeholder="costo" required name="costo"><br>
             <input type="int" placeholder="cupos" required name="cupos"><br>
+            <input type="int" placeholder="horario" required name="horario"><br>
+            <textarea  required name="comentario">descripcion </textarea><br> 
             <input type="date" placeholder="fecha curso" required name="fecha_curso"><br>
-            <!-- <input type="int" placeholder="reservas" required name="reservas"><br> -->
             <input class="boton" type="submit" name="Registrar" value="Registrar">
             <input class="boton" type="reset" name="Limpiar" value="Limpiar">
         </form>
         <input class="boton" type="submit" name="Imprimir" value="Imprimir" onclick="location.href='../controller/reportes_cursos.php'">
         <!-- <input type="submit" name="Buscar" value="Buscar" onclick="location.href='buscar.php'"> --><!--implementar lo que es buscar-->
 
-<table border="1">
+<div class="datagrid">
+<table >
 <thead>
                 <tr>
                     <th>ID</th>
                     <th>CURSOS</th>
                     <th>EXPOSITOR</th>
-                    <!-- <th>COMENTARIO</th> opcional para hacer un cambio -->
                     <th>COSTO</th>
                     <th>CUPOS</th>
                     <th>FECHA</th>
                     <th>RESERVAS</th>
+                    <th>HORARIO</th>
                     <th>ELIMINAR</th>
                     <th>MODIFICAR</th>
                 </tr>
@@ -61,22 +67,42 @@ require_once '../conexion.php';
     <tbody>
         <?php
                 if($listado->num_rows>0){
+                    $ban=0;
                     while ($row=$listado_c->fetch_assoc()){
+                        if ($ban%2==0) {
                         ?>
-                <tr>
-                    <td><?php echo $row['id_curso'];?></td>
-                    <td><?php echo $row['nombre_curso'];?></td>
-                    <td><?php echo $row['expositor'];?></td>
-                    <!-- <td><?php //echo $row['comentario'];?></td> -->
-                    <td><?php echo $row['costo'];?></td>
-                    <td><?php echo $row['cupos'];?></td>
-                    <td><?php echo $row['fecha_curso'];?></td>
-                    <td><?php echo $row['reservas'];?></td>
-                    
-                    <td><a href="../controller/crud_cursos/modificar.php? id=<?php echo $row['id_curso'];?>"><img src="../img/ico/editar.ico" alt=""></a></td>
-                    <td><a href="../controller/crud_cursos/eliminar.php? id=<?php echo $row['id_curso'];?>"><img src="../img/ico/eliminar.ico" alt=""></a></td>
-                </tr>
-                <?php
+                            <tr>
+                                <td><?php echo $row['id_curso'];?></td>
+                                <td><?php echo $row['nombre_curso'];?></td>
+                                <td><?php echo $row['expositor'];?></td>
+                                <td><?php echo $row['costo'];?></td>
+                                <td><?php echo $row['cupos'];?></td>
+                                <td><?php echo $row['fecha_curso'];?></td>
+                                <td><?php echo $row['reservas'];?></td>
+                                <td><?php echo $row['horario'];?></td>
+                                <td><a href="../controller/crud_cursos/modificar.php? id=<?php echo $row['id_curso'];?>"><img src="../img/ico/editar.ico" alt=""></a></td>
+                                <td><a href="../controller/crud_cursos/eliminar.php? id=<?php echo $row['id_curso'];?>"><img src="../img/ico/eliminar.ico" alt=""></a></td>
+                            </tr>
+                        <?php        
+                            $ban++;
+                        }
+                        else{   ?>
+                            <tr class="alt">
+                                <td><?php echo $row['id_curso'];?></td>
+                                <td><?php echo $row['nombre_curso'];?></td>
+                                <td><?php echo $row['expositor'];?></td>
+                                <td><?php echo $row['costo'];?></td>
+                                <td><?php echo $row['cupos'];?></td>
+                                <td><?php echo $row['fecha_curso'];?></td>
+                                <td><?php echo $row['reservas'];?></td>
+                                <td><?php echo $row['horario'];?></td>
+                                <td><a href="../controller/crud_cursos/modificar.php? id=<?php echo $row['id_curso'];?>"><img src="../img/ico/editar.ico" alt=""></a></td>
+                                <td><a href="../controller/crud_cursos/eliminar.php? id=<?php echo $row['id_curso'];?>"><img src="../img/ico/eliminar.ico" alt=""></a></td>
+                            </tr>
+                            
+                        <?php $ban++; }
+                
+                        
                     }
                 }  
                 else{
@@ -87,6 +113,8 @@ require_once '../conexion.php';
             </tbody>
             
         </table>
+</div>
+
             
         </body>
         </html>
